Always print ssh-proxy connection information
In order to make use of the ssh-proxy sub command, the connection
information needs to be printed, as otherwise it is not possible to know
the random port. Before this fix, the information was only ever printed,
if the `--ide` option was present. This is confusing to users, as the
docs for the `--ide` option only mention a hand full of supported IDEs.
Instead, the connection information is now always printed, unless the
quiet settings has been provided.